검색 페이지 input을 suspense에 감싸기 위한 리팩토링
